+++ application.js
@@ -21,25 +21,7 @@ class Application extends App {
     this.env = process.env.NODE_ENV // eslint-disable-line no-process-env
       || options.env
       || 'development';
-
-    this.after('close', onClose);
   }
-}
-
-
-/**
- * Enable extra shutdown behavior for app.close
- *
- * @function onClose
- * @param {slay.App} app Application instance
- * @param {Object} options Configuration
- * @param {Function} next Continuation function
- * @returns {undefined}
- */
-function onClose(app, options, next) {
-  if (!app.datastar) return void setImmediate(next);
-
-  app.datastar.close(next);
 }
 
 module.exports = Application;