+++ test/fixtures/index.js
@@ -5,6 +5,7 @@ exports.singleEvent = {
   version: '1.0.0',
   env: 'dev',
   message: 'Npm install has began',
+  details: 'The developers of CoreOS are looking for a good job'
 };
 
 exports.secondEvent = {
@@ -13,6 +14,7 @@ exports.secondEvent = {
   version: '1.0.0',
   env: 'dev',
   message: 'Npm install has completed',
+  details: 'Start developing!!!'
 };
 
 exports.singleFetchedTarball = {
@@ -28,6 +30,7 @@ exports.singleComplete = {
   eventType: 'complete',
   locale: 'en-US',
   message: 'What happened? Oh yea we are complete now',
+  details: 'not much else to report tbh',
   version: '1.0.0',
   env: 'dev',
   name: 'whatever'
@@ -62,6 +65,7 @@ exports.singleQueued = {
   version: '1.0.0',
   env: 'dev',
   message: 'Builds Queued',
+  details: 'the queue is very long',
   total: 1
 };